+++ Sources/DashX/GraphQL/Operations/Mutations/UnsubscribeContactMutation.graphql.swift
@@ -1,14 +1,14 @@
 // @generated
 // This file was automatically generated and should not be edited.
 
-@_implementationOnly import ApolloAPI
+@_exported import ApolloAPI
 
 extension DashXGql {
   class UnsubscribeContactMutation: GraphQLMutation {
     static let operationName: String = "UnsubscribeContact"
     static let operationDocument: ApolloAPI.OperationDocument = .init(
       definition: .init(
-        #"mutation UnsubscribeContact($input: UnsubscribeContactInput!) { unsubscribeContact(input: $input) { __typename id value } }"#
+        #"mutation UnsubscribeContact($input: UnsubscribeContactInput!) { unsubscribeContact(input: $input) { __typename success } }"#
       ))
 
     public var input: UnsubscribeContactInput
@@ -44,32 +44,28 @@ extension DashXGql {
 
       /// UnsubscribeContact
       ///
-      /// Parent Type: `Contact`
+      /// Parent Type: `UnsubscribeContactResponse`
       struct UnsubscribeContact: DashXGql.SelectionSet {
         let __data: DataDict
         init(_dataDict: DataDict) { __data = _dataDict }
 
-        static var __parentType: any ApolloAPI.ParentType { DashXGql.Objects.Contact }
+        static var __parentType: any ApolloAPI.ParentType { DashXGql.Objects.UnsubscribeContactResponse }
         static var __selections: [ApolloAPI.Selection] { [
           .field("__typename", String.self),
-          .field("id", DashXGql.UUID.self),
-          .field("value", String.self),
+          .field("success", Bool.self),
         ] }
         static var __fulfilledFragments: [any ApolloAPI.SelectionSet.Type] { [
           UnsubscribeContactMutation.Data.UnsubscribeContact.self
         ] }
 
-        var id: DashXGql.UUID { __data["id"] }
-        var value: String { __data["value"] }
+        var success: Bool { __data["success"] }
 
         init(
-          id: DashXGql.UUID,
-          value: String
+          success: Bool
         ) {
           self.init(unsafelyWithData: [
-            "__typename": DashXGql.Objects.Contact.typename,
-            "id": id,
-            "value": value,
+            "__typename": DashXGql.Objects.UnsubscribeContactResponse.typename,
+            "success": success,
           ])
         }
       }